Move truly common packer vars to a common location
There are a set of common Packer vars that are found in every cloud
provider packer config, all passed to Ansible, and all with the same
default value. Instead of re-defining these in each builder, define them
in a common place and include them via the Makefile.

There are a couple of limitations along with this. We can't move
variables that use functions (e.g. user or env), and we can't have vars
that reference other vars. We also can't move things that we want to
define as nil.