capi: suppress warnings on creating files without mode
Since Ansible 2.9.12 or 2.8.14, Ansible shows the following warning,
when a file gets created with `mode` not being specified in the config.

```
[WARNING]: File '/tmp/kube-apiserver.tar' created with default permissions
'600'. The previous default was '666'. Specify 'mode' to avoid this warning.
```

To get rid of such a warning, specify `mode` to correct modes like 0644
when creating files, as much as possible.

See also https://github.com/ansible/ansible/issues/71200 .