Fix Azure CentOS build
The Azure builder was not defining redhat_epel_rpm, which is why CentOS
builds were failing on package 'jq'. This moves that definition to a
common location and makes sure it is defined for all.

This also makes it such that Azure CI E2E script can test both Ubuntu
and CentOS.